feat(go-api): Integrate Go API with Python ML Inference Service
This commit finalizes the Go API service by implementing its integration with the Python ML inference service.

Key updates include:

- The `GET /api/v1/place/recommendation` endpoint is now fully functional, accepting a `user_id` and `limit`.

- The Go handler for this endpoint successfully calls the Python ML inference service's `/recommend` endpoint and processes the JSON response.

- End-to-end communication from the Go API to the Python ML service and back has been verified.

- The Go API is now a complete orchestration layer, ready to be connected to the frontend and deployed.

+++ api/internal/model/entity/places.go
@@ -9,9 +9,9 @@ type Place struct {
 	Description string  `json:"description"`
 	Category    string  `json:"category"`
 	City        string  `json:"city"`
-	Price       int     `json:"price"`
+	Price       float64 `json:"price"`
 	Rating      float64 `json:"rating"`
-	TimeMinutes int     `json:"time_minutes"`
+	TimeMinutes float64 `json:"time_minutes"`
 	Coordinate  string  `json:"coordinate"` // Stored as string, as Lat/Long are separate
 	Lat         float64 `json:"lat"`
 	Long        float64 `json:"long"`
